refactor: separate logic for heatmap
