+++ inc/classes/class-pages.php
@@ -52,6 +52,14 @@ class Pages {
 			true
 		);
 
+		wp_enqueue_script(
+			'block-analytics-script',
+			RT_TRANSCODER_URL . 'assets/build/blocks/easydam-player/analytics.js',
+			array( 'wp-element' ),
+			filemtime( RT_TRANSCODER_PATH . 'assets/build/blocks/easydam-player/analytics.js' ),
+			true
+		);
+
 		wp_localize_script(
 			'block-frontend-script',
 			'nonceData',
